Move caExport.js to export/ca.js.
